landing page complete
This was a very fun project, enjoyed the liberties I felt I could take and the freedom I had to do what I wanted with the page.
I'm unsure if it's my implementation or if I just don't like the color palette I've selected. Might revisit this in the future to see what more I can do to enhance the page overall.